Fixed warnings and updated to recommended project settings

+++ DMTabBar/DMTabBar/DMAppDelegate.m
@@ -35,7 +35,7 @@
         
         DMTabBarItem *item1 = [DMTabBarItem tabBarItemWithIcon:iconImage tag:idx];
         item1.toolTip = [objDict objectForKey:@"title"];
-        item1.keyEquivalent = [NSString stringWithFormat:@"%d",idx];
+        item1.keyEquivalent = [NSString stringWithFormat:@"%lu",(unsigned long)idx];
         item1.keyEquivalentModifierMask = NSCommandKeyMask;
         [items addObject:item1];
     }];
